+++ source/jdbcDriverOOo/source/io/github/prrvchr/driver/provider/DriverManagerHelper.java
@@ -39,16 +39,11 @@ import java.sql.Driver;
 import java.sql.DriverManager;
 import java.util.ArrayList;
 import java.util.Enumeration;
-import java.util.Iterator;
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.ServiceConfigurationError;
-import java.util.ServiceLoader;
 
 import com.sun.star.beans.PropertyValue;
 import com.sun.star.container.XHierarchicalNameAccess;
 import com.sun.star.lib.util.StringHelper;
-import com.sun.star.logging.LogLevel;
 import com.sun.star.sdbc.SQLException;
 import com.sun.star.uno.UnoRuntime;
 import com.sun.star.uno.XComponentContext;
@@ -57,7 +52,6 @@ import com.sun.star.util.XMacroExpander;
 
 import io.github.prrvchr.driver.helper.DBException;
 import io.github.prrvchr.uno.agent.UnoAgent;
-import io.github.prrvchr.uno.helper.ResourceBasedEventLogger;
 import io.github.prrvchr.uno.helper.SharedResources;
 import io.github.prrvchr.uno.helper.UnoHelper;
 import io.github.prrvchr.uno.logger.UnoLoggerFinder;
@@ -65,10 +59,13 @@ import io.github.prrvchr.uno.logger.UnoLoggerFinder;
 
 public class DriverManagerHelper {
 
+    private static final String DRIVER_NAME = "DriverTypeDisplayName";
     private static final String DRIVER_CLASS = "JavaDriverClass";
     private static final String DRIVER_CLASS_PATH = "JavaDriverClassPath";
     private static final String PLUGIN_JAR_PATH = "/plugin/UnoLogger.jar";
     private static final String JAR = "jar";
+    private static final String JAR_URL_PREFIX = JAR + ":";
+    private static final String JAR_URL_SUFFIX = "!/";
     private static final String DOT = ".";
     private static final String EXPAND_PROTOCOL = "vnd.sun.star.expand:";
 
@@ -100,38 +97,71 @@ public class DriverManagerHelper {
         return registered;
     }
 
-    public static final void registerDriver(final XComponentContext context,
-                                            final XInterface source,
-                                            final XHierarchicalNameAccess config,
-                                            final ResourceBasedEventLogger logger,
-                                            final String subProtocol,
-                                            final PropertyValue[] info)
-        throws SQLException {
-        System.out.println("DriverManagerHelper.registerDriver() 1");
-        final String clazz = (String) DriverPropertiesHelper.getConfigProperties(config, info, subProtocol,
-                                                                                 DRIVER_CLASS, null);
-        System.out.println("DriverManagerHelper.registerDriver() 2 Class: " + clazz);
-        final File[] files = getDriverClassFiles(context, source, config, logger, subProtocol, info, clazz);
-        System.out.println("DriverManagerHelper.registerDriver() 3 Files length: " + files.length);
-        registerJdbcDriver(source, logger, files, clazz);
+    public static final String getDriverName(final XHierarchicalNameAccess config, final String subProtocol) {
+        return (String) DriverPropertiesHelper.getConfig(config, subProtocol, DRIVER_NAME, null);
     }
 
-    private static final File[] getDriverClassFiles(final XComponentContext context,
-                                                    final XInterface source,
-                                                    final XHierarchicalNameAccess config,
-                                                    final ResourceBasedEventLogger logger,
-                                                    final String protocol,
-                                                    final PropertyValue[] info,
-                                                    final String clazz)
+    public static final String getDriverClassPath(final XComponentContext context,
+                                                  final XInterface source,
+                                                  final XHierarchicalNameAccess config,
+                                                  final PropertyValue[] info,
+                                                  final String subProtocol,
+                                                  final String name)
         throws SQLException {
-        System.out.println("DriverManagerHelper.getDriverClassFiles() 1");
-        File[] files = new File[0];
-        Object path = DriverPropertiesHelper.getConfigProperties(config, info, protocol, DRIVER_CLASS_PATH, null);
-        if (clazz != null && path != null) {
-            System.out.println("DriverManagerHelper.getDriverClassFiles() 2");
-            files = getDriverArchiveFiles(source, logger, expandURL(context, source, (String) path), clazz);
+        String path = (String) DriverPropertiesHelper.getConfigProperties(config, info, subProtocol,
+                                                                          DRIVER_CLASS_PATH, null);
+        String url = null;
+        if (path != null && !path.isBlank()) {
+            url = expandURL(context, source, path);
         }
-        return files;
+        if (url == null) {
+            final int resource = Resources.STR_LOG_DRIVER_CLASS_NOT_FOUND;
+            final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, name, subProtocol,
+                                                                                         DRIVER_CLASS_PATH);
+            throw DBException.getSQLException(msg, source, StandardSQLState.SQL_UNABLE_TO_CONNECT);
+        }
+        return url;
+    }
+
+    public static final Driver getJdbcDriver(final XInterface source,
+                                             final XHierarchicalNameAccess config,
+                                             final String subProtocol,
+                                             final PropertyValue[] info,
+                                             final String url,
+                                             final String name)
+        throws SQLException {
+        final String clazz = (String) DriverPropertiesHelper.getConfigProperties(config, info, subProtocol,
+                                                                                 DRIVER_CLASS, null);
+        if (clazz == null || clazz.isBlank()) {
+            final int resource = Resources.STR_LOG_DRIVER_CLASS_NOT_FOUND;
+            final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, name, subProtocol,
+                                                                                         DRIVER_CLASS);
+            throw DBException.getSQLException(msg, source, StandardSQLState.SQL_UNABLE_TO_CONNECT);
+        }
+        return getDriverByClassName(source, url, clazz);
+    }
+
+    public static final void registerDriver(final XInterface source,
+                                            final Driver driver,
+                                            final String url,
+                                            final String path,
+                                            final String name)
+        throws SQLException {
+        URL location = null;
+        try {
+            location = new URL(path);
+        } catch (MalformedURLException e) { }
+        // XXX: Does Java provide the jdbcDriverOOo embedded driver?
+        if (location == null || !url.startsWith(location.getPath())) {
+            final int resource = Resources.STR_LOG_DRIVER_JAVA_CLASS_NOT_SUPPORTED;
+            final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, url, name);
+            throw DBException.getSQLException(msg, source, StandardSQLState.SQL_UNABLE_TO_CONNECT);
+        }
+        try {
+            DriverManager.registerDriver(new DriverWrapper(driver));
+        } catch (java.sql.SQLException e) {
+            throw DBException.getSQLException(e.getMessage(), source, StandardSQLState.SQL_UNABLE_TO_CONNECT);
+        }
     }
 
     private static final String expandURL(final XComponentContext context,
@@ -139,7 +169,7 @@ public class DriverManagerHelper {
                                           final String url)
         throws SQLException {
         String expanded = url;
-        if (url != null && url.startsWith(EXPAND_PROTOCOL)) {
+        if (url.startsWith(EXPAND_PROTOCOL)) {
             try {
                 final Object service = context.getValueByName("/singletons/com.sun.star.util.theMacroExpander");
                 final XMacroExpander expander = UnoRuntime.queryInterface(XMacroExpander.class, service);
@@ -149,131 +179,30 @@ public class DriverManagerHelper {
                 // expand macro string
                 expanded = expander.expandMacros(macro);
             } catch (UnsupportedEncodingException e) {
-                SQLException ex = new SQLException(e, e.getMessage());
-                ex.Context = source;
-                throw ex;
+                throw DBException.getSQLException(e.getMessage(), source, StandardSQLState.SQL_UNABLE_TO_CONNECT);
             }
         }
         return expanded;
     }
 
-    private static final File[] getDriverArchiveFiles(final XInterface source,
-                                                      final ResourceBasedEventLogger logger,
-                                                      final String location,
-                                                      final String clazz) throws SQLException {
-        // XXX: In order to allow the loading of drivers requiring several Java archives,
-        // XXX: the JavaDriverClassPath parameter can be a file or a folder
-        File file = getArchiveFile(source, location, clazz);
-        final List<File> files = new ArrayList<>();
-        if (file.isDirectory()) {
-            for (final File f : file.listFiles()) {
-                if (!f.isDirectory() && isArchiveFile(f)) {
-                    System.out.println("DriverManagerHelper.getDriverArchiveFiles() 1 File: " + f.getName());
-                    files.add(f);
-                }
-            }
-        } else if (isArchiveFile(file)) {
-            System.out.println("DriverManagerHelper.getDriverArchiveFiles() 2 File: " + file.getName());
-            files.add(file);
-        }
-
-        // XXX: JavaDriverClassPath must contain at least one archive
-        if (files.isEmpty()) {
-            final int resource;
-            if (file.isDirectory()) {
-                resource = Resources.STR_LOG_DRIVER_CLASS_PATH_EMPTY;
-            } else {
-                resource = Resources.STR_LOG_DRIVER_CLASS_PATH_NO_ARCHIVE;
-            }
-            final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource,
-                                                          location, clazz, DRIVER_CLASS_PATH);
-            throw DBException.getSQLException(msg, source, StandardSQLState.SQL_UNABLE_TO_CONNECT);
-        }
-        return files.toArray(new File[files.size()]);
-    }
-
-    private static final File getArchiveFile(final XInterface source,
-                                             final String location,
-                                             final String clazz) throws SQLException {
-        File file = null;
-        try {
-            file = new File(new URI(location).normalize());
-        } catch (URISyntaxException e) {
-            throw getClassPathParseError(source, e, location, clazz);
-        }
-        if (file == null || !file.exists()) {
-            final int resource = Resources.STR_LOG_DRIVER_CLASS_PATH_NOT_FOUND;
-            final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, location,
-                                                                                         clazz, DRIVER_CLASS_PATH);
-            throw DBException.getSQLException(msg, source, StandardSQLState.SQL_UNABLE_TO_CONNECT);
-        }
-        return file;
-    }
-
-    private static final void registerJdbcDriver(final XInterface source,
-                                                 final ResourceBasedEventLogger logger,
-                                                 final File[] files,
-                                                 final String name)
-        throws SQLException {
-        java.sql.Driver driver = null;
-        try {
-            System.out.println("DriverManagerHelper.registerJdbcDriver() 1");
-            driver = getDriverByServiceLoader(source, logger, files, name);
-            if (driver == null) {
-                System.out.println("DriverManagerHelper.registerJdbcDriver() 2");
-                driver = getDriverByClassName(source, logger, files, name);
-                System.out.println("DriverManagerHelper.registerJdbcDriver() 3");
-            }
-            System.out.println("DriverManagerHelper.registerJdbcDriver() 4 Driver: " + driver.getMajorVersion());
-            if (driver != null) {
-                System.out.println("DriverManagerHelper.registerJdbcDriver() 5");
-                DriverManager.registerDriver(driver);
-                System.out.println("DriverManagerHelper.registerJdbcDriver() 6");
-            }
-        } catch (java.sql.SQLException e) {
-            final int resource = Resources.STR_LOG_DRIVER_REGISTER_ERROR;
-            final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, name);
-            throw DBException.getSQLException(msg, StandardSQLState.SQL_UNABLE_TO_CONNECT,
-                                              DBException.getSQLException(source, e));
-        }
-    }
-
-    private static final java.sql.Driver getDriverByServiceLoader(final XInterface source,
-                                                                  final ResourceBasedEventLogger logger,
-                                                                  final File[] files,
-                                                                  final String name) {
-        java.sql.Driver driver = null;
-        Iterator<java.sql.Driver> drivers = null;
-        try {
-            URL[] urls = getDriverArchiveUrls(source, logger, files, name, "", "");
-            drivers = ServiceLoader.load(java.sql.Driver.class, new URLClassLoader(urls)).iterator();
-            if (drivers != null && drivers.hasNext()) {
-                driver = new DriverWrapper(drivers.next());
-            }
-        } catch (SQLException | ServiceConfigurationError | SecurityException | NoSuchElementException e) {
-            logger.logprb(LogLevel.SEVERE, Resources.STR_LOG_DRIVER_SERVICE_LOADER_ERROR, name);
-        }
-        return driver;
-    }
-
     private static final java.sql.Driver getDriverByClassName(final XInterface source,
-                                                              final ResourceBasedEventLogger logger,
-                                                              final File[] files,
+                                                              final String url,
                                                               final String name) throws SQLException {
         java.sql.Driver driver = null;
         try {
+            final File[] files = getDriverArchiveFiles(source, url, name);
             // XXX: Pick your JDBC driver at runtime: https://www.kfu.com/~nsayer/Java/dyn-jdbc.html
-            URL[] urls = getDriverArchiveUrls(source, logger, files, name, JAR + ":", "!/");
+            URL[] urls = getDriverArchiveUrls(source, files, name);
             final URLClassLoader loader = new URLClassLoader(urls, source.getClass().getClassLoader());
             final Class<?> clazz = Class.forName(name, true, loader);
-            driver = new DriverWrapper((java.sql.Driver) clazz.getDeclaredConstructor().newInstance());
+            driver = (java.sql.Driver) clazz.getDeclaredConstructor().newInstance();
         } catch (UnsupportedClassVersionError e) {
             final String version = System.getProperty("java.version");
             final int resource = Resources.STR_LOG_DRIVER_UNSUPPORTED_JAVA_VERSION;
             final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, version, name);
             throw DBException.getSQLException(msg, source, StandardSQLState.SQL_UNABLE_TO_CONNECT, e);
         } catch (ClassNotFoundException | IllegalAccessException e) {
-            final int resource = Resources.STR_LOG_DRIVER_JAVA_CLASS_NOT_FOUND;
+            final int resource = Resources.STR_LOG_DRIVER_CLASS_NOT_FOUND;
             final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, name,
                                                                                          DRIVER_CLASS_PATH,
                                                                                          DRIVER_CLASS);
@@ -286,25 +215,56 @@ public class DriverManagerHelper {
         return driver;
     }
 
+    private static final File[] getDriverArchiveFiles(final XInterface source,
+                                                      final String location,
+                                                      final String clazz) throws SQLException {
+        File file = null;
+        try {
+            file = new File(new URI(location).normalize());
+        } catch (URISyntaxException e) { }
+        List<File> files = null;
+        if (file != null && file.exists()) {
+            files = getDriverArchiveFiles(file);
+        }
+        // XXX: JavaDriverClassPath must contain at least one archive
+        if (files == null || files.isEmpty()) {
+            final int resource = getDriverArchiveFilesResource(file);
+            final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, location,
+                                                                                         clazz, DRIVER_CLASS_PATH);
+            throw DBException.getSQLException(msg, source, StandardSQLState.SQL_UNABLE_TO_CONNECT);
+        }
+        return files.toArray(new File[files.size()]);
+    }
+
+    private static final List<File> getDriverArchiveFiles(final File file) {
+        // XXX: In order to allow the loading of drivers requiring several Java archives,
+        // XXX: the JavaDriverClassPath parameter can be a file or a folder
+        final List<File> files = new ArrayList<>();
+        if (file.isDirectory()) {
+            for (final File f : file.listFiles()) {
+                if (!f.isDirectory() && isArchiveFile(f)) {
+                    files.add(f);
+                }
+            }
+        } else if (isArchiveFile(file)) {
+            files.add(file);
+        }
+        return files;
+    }
+
     private static final boolean isArchiveFile(final File file) {
         return file.getName().toLowerCase().endsWith(DOT + JAR);
     }
 
     private static final URL[] getDriverArchiveUrls(final XInterface source,
-                                                    final ResourceBasedEventLogger logger,
                                                     final File[] files,
-                                                    final String clazz,
-                                                    String prefix,
-                                                    String suffix) throws SQLException {
+                                                    final String clazz) throws SQLException {
         String url = "";
         final List<URL> urls = new ArrayList<>();
         try {
             for (File file : files) {
-                url = prefix + Path.of(file.getAbsolutePath()).toUri() + suffix;
-                System.out.println("DriverManagerHelper.getDriverArchiveUrls() 1 url: " + url);
+                url = JAR_URL_PREFIX + Path.of(file.getAbsolutePath()).toUri() + JAR_URL_SUFFIX;
                 urls.add(new URL(url));
-                System.out.println("DriverManagerHelper.getDriverArchiveUrls() 2 url: " + url);
-                logger.logprb(LogLevel.INFO, Resources.STR_LOG_DRIVER_ARCHIVE_LOADING, url);
             }
         } catch (MalformedURLException e) {
             throw getClassPathParseError(source, e, url, clazz);
@@ -320,6 +280,16 @@ public class DriverManagerHelper {
         final String msg = SharedResources.getInstance().getResourceWithSubstitution(resource, url,
                                                         clazz, DRIVER_CLASS_PATH);
         return DBException.getSQLException(msg, source, StandardSQLState.SQL_UNABLE_TO_CONNECT, e);
+    }
+
+    private static final int getDriverArchiveFilesResource(final File file) {
+        final int resource;
+        if (file.isDirectory()) {
+            resource = Resources.STR_LOG_DRIVER_CLASS_PATH_EMPTY;
+        } else {
+            resource = Resources.STR_LOG_DRIVER_CLASS_PATH_NO_ARCHIVE;
+        }
+        return resource;
     }
 
 }